requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.11.2
blinker==1.7.0
cachetools==5.3.3
certifi==2023.11.17
charset-normalizer==2.1.1
click==8.1.7
exceptiongroup==1.2.0
Flask==2.2.5
frozenlist==1.4.1
future==1.0.0
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-auth==2.29.0
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
line-bot-sdk==2.4.1
MarkupSafe==2.1.3
multidict==6.0.5
opencc-python-reimplemented==0.1.4
outcome==1.3.0.post0
packaging==23.2
playwright==1.42.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.6.4
pydantic_core==2.16.3
pyee==11.0.1
pyperclip==1.8.2
PySocks==1.7.1
python-dotenv==0.21.1
requests==2.28.1
rsa==4.9
selenium==4.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tqdm==4.66.2
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==1.26.18
webdriver-manager==4.0.1
Werkzeug==2.3.8
wsproto==1.2.0
yarl==1.9.4
youtube-transcript-api==0.5.0
zipp==3.17.0
